Object materials: more robust encyclopedia and lookup
This is adapted from a pull request against SpliceHack by Michael Meyer,
in which the parser, if it detects a material prefix, will check for
both the string containing the prefix and the string not containing the
prefix. Without this, attempting to look up something like "silver
banded mail" will fail because it doesn't recognize that as the base
name for an object.

Note: it did work properly if you looked up an item from an inventory
window, because that goes through a simple_typename conversion, but
actually searching for the string of the item did not work.