Print headings for monster and object lookup and encyclopedia entry
This is in order to make it clear to the player what they've asked to
see more information about, in case it's in any way unclear.

For the lookups, this will render a heading in bold of the matched
object or monster, then a blank line, then the lookup body (which hasn't
been changed). Additionally, monster lookup printed everything in bold
whereas object lookup did not, so I made the monster lookup body
non-bold.

This refactors the logic of the string matching somewhat to store a
pointer to the string that matched something in the database, and the
string in data.base it matched against. Both of these are printed in the
heading for the encyclopedia entry. I debated whether I should print the
latter string, but figured it was useful if the player looks up "ring of
protection" and gets an encyclopedia entry that matches "ring*" (and
therefore the entry isn't expected to be specifically about rings of
protection).